refactor: move flavours into enum
