+++ rust/src/lib.rs
@@ -12,7 +12,7 @@
 //! # Example
 //!
 //! ```rust
-//! use catppuccin::MOCHA;
+//! use catppuccin::Flavour;
 //!
 //! struct Button {
 //!     text: String,
@@ -22,7 +22,7 @@
 //! fn confirm(text: String) -> Button {
 //!     Button {
 //!         text,
-//!         background_colour: MOCHA.green.hex(),
+//!         background_colour: Flavour::Mocha.palette().green.hex(),
 //!     }
 //! }
 //! ```
@@ -55,38 +55,29 @@
 mod colour;
 pub use colour::Colour;
 
+mod palette;
+pub use palette::Palette;
+
 mod flavour;
 pub use flavour::Flavour;
 
-mod latte;
-pub use latte::LATTE;
-
-mod frappe;
-pub use frappe::FRAPPE;
-
-mod macchiato;
-pub use macchiato::MACCHIATO;
-
-mod mocha;
-pub use mocha::MOCHA;
-
 /// Returns an iterator over the four delicious Catppuccin flavours.
 pub fn flavours_iter() -> std::array::IntoIter<Flavour, 4> {
-    [LATTE, FRAPPE, MACCHIATO, MOCHA].into_iter()
+    [Flavour::Latte, Flavour::Frappe, Flavour::Macchiato, Flavour::Mocha].into_iter()
 }
 
 #[cfg(test)]
 mod tests {
-    use super::{flavours_iter, FRAPPE, LATTE, MACCHIATO, MOCHA};
+    use super::{flavours_iter, Flavour};
 
     /// Ensures flavours are iterated in the correct order.
     #[test]
     fn test_flavours_iter() {
         let mut flavours = flavours_iter();
-        assert_eq!(flavours.next(), Some(LATTE));
-        assert_eq!(flavours.next(), Some(FRAPPE));
-        assert_eq!(flavours.next(), Some(MACCHIATO));
-        assert_eq!(flavours.next(), Some(MOCHA));
+        assert_eq!(flavours.next(), Some(Flavour::Latte));
+        assert_eq!(flavours.next(), Some(Flavour::Frappe));
+        assert_eq!(flavours.next(), Some(Flavour::Macchiato));
+        assert_eq!(flavours.next(), Some(Flavour::Mocha));
         assert_eq!(flavours.next(), None);
     }
 
@@ -94,33 +85,34 @@ mod tests {
     #[test]
     fn test_colours_iter() {
         for flavour in flavours_iter() {
-            let mut colours = flavour.iter();
-            assert_eq!(colours.next(), Some(&flavour.rosewater));
-            assert_eq!(colours.next(), Some(&flavour.flamingo));
-            assert_eq!(colours.next(), Some(&flavour.pink));
-            assert_eq!(colours.next(), Some(&flavour.mauve));
-            assert_eq!(colours.next(), Some(&flavour.red));
-            assert_eq!(colours.next(), Some(&flavour.maroon));
-            assert_eq!(colours.next(), Some(&flavour.peach));
-            assert_eq!(colours.next(), Some(&flavour.yellow));
-            assert_eq!(colours.next(), Some(&flavour.green));
-            assert_eq!(colours.next(), Some(&flavour.teal));
-            assert_eq!(colours.next(), Some(&flavour.sky));
-            assert_eq!(colours.next(), Some(&flavour.sapphire));
-            assert_eq!(colours.next(), Some(&flavour.blue));
-            assert_eq!(colours.next(), Some(&flavour.lavender));
-            assert_eq!(colours.next(), Some(&flavour.text));
-            assert_eq!(colours.next(), Some(&flavour.subtext1));
-            assert_eq!(colours.next(), Some(&flavour.subtext0));
-            assert_eq!(colours.next(), Some(&flavour.overlay2));
-            assert_eq!(colours.next(), Some(&flavour.overlay1));
-            assert_eq!(colours.next(), Some(&flavour.overlay0));
-            assert_eq!(colours.next(), Some(&flavour.surface2));
-            assert_eq!(colours.next(), Some(&flavour.surface1));
-            assert_eq!(colours.next(), Some(&flavour.surface0));
-            assert_eq!(colours.next(), Some(&flavour.base));
-            assert_eq!(colours.next(), Some(&flavour.mantle));
-            assert_eq!(colours.next(), Some(&flavour.crust));
+            let palette = flavour.palette();
+            let mut colours = palette.iter();
+            assert_eq!(colours.next(), Some(&palette.rosewater));
+            assert_eq!(colours.next(), Some(&palette.flamingo));
+            assert_eq!(colours.next(), Some(&palette.pink));
+            assert_eq!(colours.next(), Some(&palette.mauve));
+            assert_eq!(colours.next(), Some(&palette.red));
+            assert_eq!(colours.next(), Some(&palette.maroon));
+            assert_eq!(colours.next(), Some(&palette.peach));
+            assert_eq!(colours.next(), Some(&palette.yellow));
+            assert_eq!(colours.next(), Some(&palette.green));
+            assert_eq!(colours.next(), Some(&palette.teal));
+            assert_eq!(colours.next(), Some(&palette.sky));
+            assert_eq!(colours.next(), Some(&palette.sapphire));
+            assert_eq!(colours.next(), Some(&palette.blue));
+            assert_eq!(colours.next(), Some(&palette.lavender));
+            assert_eq!(colours.next(), Some(&palette.text));
+            assert_eq!(colours.next(), Some(&palette.subtext1));
+            assert_eq!(colours.next(), Some(&palette.subtext0));
+            assert_eq!(colours.next(), Some(&palette.overlay2));
+            assert_eq!(colours.next(), Some(&palette.overlay1));
+            assert_eq!(colours.next(), Some(&palette.overlay0));
+            assert_eq!(colours.next(), Some(&palette.surface2));
+            assert_eq!(colours.next(), Some(&palette.surface1));
+            assert_eq!(colours.next(), Some(&palette.surface0));
+            assert_eq!(colours.next(), Some(&palette.base));
+            assert_eq!(colours.next(), Some(&palette.mantle));
+            assert_eq!(colours.next(), Some(&palette.crust));
         }
     }
 }
